@@ -1,8 +1,8 @@
 {
   "_from": "extend@~3.0.1",
-  "_id": "extend@3.0.1",
+  "_id": "extend@3.0.2",
   "_inBundle": false,
-  "_integrity": "sha1-p1Xqe8Gt/MWjHOfnYtuq3F5jZEQ=",
+  "_integrity": "sha512-fjquC59cD7CyW6urNXK0FBufkZcoiGG80wTuPujX590cB5Ttln20E2UB4S/WARVqhXffZl2LNgS+gQdPIIim/g==",
   "_location": "/extend",
   "_phantomChildren": {},
   "_requested": {
@@ -18,10 +18,10 @@
   "_requiredBy": [
     "/postman-request"
   ],
-  "_resolved": "https://registry.npmjs.org/extend/-/extend-3.0.1.tgz",
-  "_shasum": "a755ea7bc1adfcc5a31ce7e762dbaadc5e636444",
+  "_resolved": "https://registry.npmjs.org/extend/-/extend-3.0.2.tgz",
+  "_shasum": "f8b1136b4071fbd8eb140aff858b1019ec2915fa",
   "_spec": "extend@~3.0.1",
-  "_where": "/var/lib/buildkite-agent/builds/postman-ci-agent-linux-1/postman/electron-package-and-deploy/build/node_modules/postman-request",
+  "_where": "/var/lib/buildkite-agent/builds/postman-ci-agent-linux-8/postman/app-package-and-deploy/build/node_modules/postman-request",
   "author": {
     "name": "Stefan Thomas",
     "email": "justmoon@members.fsf.org",
@@ -41,11 +41,11 @@
   "deprecated": false,
   "description": "Port of jQuery.extend for node.js and the browser",
   "devDependencies": {
-    "@ljharb/eslint-config": "^11.0.0",
+    "@ljharb/eslint-config": "^12.2.1",
     "covert": "^1.1.0",
-    "eslint": "^3.19.0",
+    "eslint": "^4.19.1",
     "jscs": "^3.0.7",
-    "tape": "^4.6.3"
+    "tape": "^4.9.1"
   },
   "homepage": "https://github.com/justmoon/node-extend#readme",
   "keywords": [
@@ -71,5 +71,5 @@
     "test": "npm run tests-only",
     "tests-only": "node test"
   },
-  "version": "3.0.1"
+  "version": "3.0.2"
 }
