@@ -1,32 +1,32 @@
 {
-  "_from": "uuid@3.2.1",
-  "_id": "uuid@3.2.1",
+  "_from": "uuid@3.3.2",
+  "_id": "uuid@3.3.2",
   "_inBundle": false,
-  "_integrity": "sha512-jZnMwlb9Iku/O3smGWvZhauCf6cvvpKi4BKRiliS3cxnI+Gz9j5MEpTz2UFuXiKPJocb7gnsLHwiS05ige5BEA==",
+  "_integrity": "sha512-yXJmeNaw3DnnKAOKJE51sL/ZaYfWJRl1pK9dr19YFCu0ObS231AB1/LbqTKRAQ5kw8A90rA6fr4riOUpTZvQZA==",
   "_location": "/uuid",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "uuid@3.2.1",
+    "raw": "uuid@3.3.2",
     "name": "uuid",
     "escapedName": "uuid",
-    "rawSpec": "3.2.1",
+    "rawSpec": "3.3.2",
     "saveSpec": null,
-    "fetchSpec": "3.2.1"
+    "fetchSpec": "3.3.2"
   },
   "_requiredBy": [
     "/",
     "/postman-collection",
     "/postman-request",
     "/postman-runtime",
-    "/postman-sandbox",
-    "/uvm"
+    "/postman-runtime/serialised-error",
+    "/postman-sandbox"
   ],
-  "_resolved": "https://registry.npmjs.org/uuid/-/uuid-3.2.1.tgz",
-  "_shasum": "12c528bb9d58d0b9265d9a2f6f0fe8be17ff1f14",
-  "_spec": "uuid@3.2.1",
-  "_where": "/var/lib/buildkite-agent/builds/postman-ci-agent-linux-1/postman/electron-package-and-deploy/build",
+  "_resolved": "https://registry.npmjs.org/uuid/-/uuid-3.3.2.tgz",
+  "_shasum": "1b4af4955eb3077c501c23872fc6513811587131",
+  "_spec": "uuid@3.3.2",
+  "_where": "/var/lib/buildkite-agent/builds/postman-ci-agent-linux-8/postman/app-package-and-deploy/build",
   "bin": {
     "uuid": "./bin/uuid"
   },
@@ -39,6 +39,11 @@
     "url": "https://github.com/kelektiv/node-uuid/issues"
   },
   "bundleDependencies": false,
+  "commitlint": {
+    "extends": [
+      "@commitlint/config-conventional"
+    ]
+  },
   "contributors": [
     {
       "name": "Robert Kieffer",
@@ -61,14 +66,16 @@
       "email": "shtylman@gmail.com"
     }
   ],
-  "dependencies": {},
   "deprecated": false,
   "description": "RFC4122 (v1, v4, and v5) UUIDs",
   "devDependencies": {
-    "eslint": "4.5.0",
-    "mocha": "3.1.2",
+    "@commitlint/cli": "7.0.0",
+    "@commitlint/config-conventional": "7.0.1",
+    "eslint": "4.19.1",
+    "husky": "0.14.3",
+    "mocha": "5.2.0",
     "runmd": "1.0.1",
-    "standard-version": "4.2.0"
+    "standard-version": "4.4.0"
   },
   "homepage": "https://github.com/kelektiv/node-uuid#readme",
   "keywords": [
@@ -83,10 +90,11 @@
     "url": "git+https://github.com/kelektiv/node-uuid.git"
   },
   "scripts": {
+    "commitmsg": "commitlint -E GIT_PARAMS",
     "md": "runmd --watch --output=README.md README_js.md",
     "prepare": "runmd --output=README.md README_js.md",
     "release": "standard-version",
     "test": "mocha test/test.js"
   },
-  "version": "3.2.1"
+  "version": "3.3.2"
 }
