@@ -1,6 +1,6 @@
 webpackJsonp([17],{
 
-/***/ 1641:
+/***/ 1783:
 /***/ (function(module, exports) {
 
 module.exports.id = 'ace/mode/xml_worker';
@@ -8,7 +8,7 @@ module.exports.src = "\"no use strict\";(function(window){function resolveModule
 
 /***/ }),
 
-/***/ 2146:
+/***/ 2278:
 /***/ (function(module, exports, __webpack_require__) {
 
 ace.define("ace/mode/xml_highlight_rules",["require","exports","module","ace/lib/oop","ace/mode/text_highlight_rules"], function(acequire, exports, module) {
@@ -657,7 +657,7 @@ oop.inherits(Mode, TextMode);
     this.blockComment = {start: "<!--", end: "-->"};
 
     this.createWorker = function(session) {
-        var worker = new WorkerClient(["ace"], __webpack_require__(1641), "Worker");
+        var worker = new WorkerClient(["ace"], __webpack_require__(1783), "Worker");
         worker.attachToDocument(session.getDocument());
 
         worker.on("error", function(e) {
